rework of entire site + added media, widgets, and gallery
remake of the entire site with less messy code and added media, widget, and css image gallery. Added java script to play page turn sound when description widget is clicked and an error sound for when the review form is not completed. Gallery is used for the multiple books on the pages where they are showcased. Changed /image to /media and added js file.